feat: revamp hide on state change plugin

+++ src/utils/state.rs
@@ -1,26 +1,11 @@
 use crate::prelude::{tween::DespawnOnTweenCompleted, *};
 
 #[derive(Default)]
-pub struct HideOnStatePlugin<TState: States> {
-    pub restore_on_enter_states: Vec<TState>,
-    pub restore_on_exit_states: Vec<TState>,
-    pub hide_on_enter_states: Vec<TState>,
-    pub hide_on_exit_states: Vec<TState>,
-}
+pub struct HideOnStatePlugin<TState: States>(PhantomData<TState>);
 impl<TState: States> Plugin for HideOnStatePlugin<TState> {
     fn build(&self, app: &mut App) {
-        for state in &self.restore_on_enter_states {
-            app.add_systems(OnEnter(state.clone()), restore::<TState>);
-        }
-        for state in &self.restore_on_exit_states {
-            app.add_systems(OnExit(state.clone()), restore::<TState>);
-        }
-        for state in &self.hide_on_enter_states {
-            app.add_systems(OnEnter(state.clone()), hide::<TState>);
-        }
-        for state in &self.hide_on_exit_states {
-            app.add_systems(OnExit(state.clone()), hide::<TState>);
-        }
+        app.add_systems(Update, restore::<TState>.run_if(state_changed::<TState>));
+        app.add_systems(Update, hide::<TState>.run_if(state_changed::<TState>));
     }
 }
 
@@ -40,28 +25,102 @@ pub enum HideTween {
 pub struct HideOnStateChange<TState: States> {
     tween: HideTween,
     despawn: bool,
-    _state: PhantomData<TState>,
+    restore_on_enter_states: Vec<TState>,
+    restore_on_exit_states: Vec<TState>,
+    hide_on_enter_states: Vec<TState>,
+    hide_on_exit_states: Vec<TState>,
 }
+#[allow(dead_code)]
 impl<TState: States> HideOnStateChange<TState> {
-    pub fn new(tween: HideTween) -> Self {
-        Self {
-            tween,
-            despawn: false,
-            _state: PhantomData,
-        }
+    pub fn restore_on_enter(tween: HideTween, state: TState) -> Self {
+        let mut hide = Self::base(tween);
+        hide.restore_on_enter_states = vec![state];
+        hide
+    }
+
+    pub fn with_restore_on_enter(mut self, state: TState) -> Self {
+        self.restore_on_enter_states.push(state);
+        self
+    }
+
+    pub fn restore_on_exit(tween: HideTween, state: TState) -> Self {
+        let mut hide = Self::base(tween);
+        hide.restore_on_enter_states = vec![state];
+        hide
+    }
+
+    pub fn with_restore_on_exit(mut self, state: TState) -> Self {
+        self.restore_on_exit_states.push(state);
+        self
+    }
+
+    pub fn hide_on_enter(tween: HideTween, state: TState) -> Self {
+        let mut hide = Self::base(tween);
+        hide.hide_on_enter_states = vec![state];
+        hide
+    }
+
+    pub fn with_hide_on_enter(mut self, state: TState) -> Self {
+        self.hide_on_enter_states.push(state);
+        self
+    }
+
+    pub fn hide_on_exit(tween: HideTween, state: TState) -> Self {
+        let mut hide = Self::base(tween);
+        hide.hide_on_exit_states = vec![state];
+        hide
+    }
+
+    pub fn with_hide_on_exit(mut self, state: TState) -> Self {
+        self.hide_on_exit_states.push(state);
+        self
     }
 
     pub fn with_despawn(mut self) -> Self {
         self.despawn = true;
         self
     }
+
+    fn base(tween: HideTween) -> Self {
+        Self {
+            tween,
+            despawn: false,
+            restore_on_enter_states: Vec::default(),
+            restore_on_exit_states: Vec::default(),
+            hide_on_enter_states: Vec::default(),
+            hide_on_exit_states: Vec::default(),
+        }
+    }
 }
 
 fn hide<TState: States>(
     mut cmd: Commands,
     hide_q: Query<(Entity, &HideOnStateChange<TState>, &Transform)>,
+    mut state_reader: MessageReader<StateTransitionEvent<TState>>,
 ) {
+    let msgs: Vec<_> = state_reader.read().collect();
+    if msgs.is_empty() {
+        return;
+    }
+
     for (e, hide, hide_t) in hide_q {
+        if !msgs.iter().any(|msg| {
+            if let Some(entered) = msg.entered.as_ref()
+                && hide.hide_on_enter_states.contains(entered)
+            {
+                return true;
+            }
+            if let Some(exited) = msg.exited.as_ref()
+                && hide.hide_on_exit_states.contains(exited)
+            {
+                return true;
+            }
+
+            false
+        }) {
+            return;
+        }
+
         let pos = hide_t.translation;
         let new_pos = match hide.tween {
             HideTween::AbsoluteX(x) => pos.with_x(x),
@@ -82,9 +141,36 @@ fn hide<TState: States>(
 
 fn restore<TState: States>(
     mut cmd: Commands,
-    restore_q: Query<(Entity, &TweenBackOnStateChange<TState>)>,
+    restore_q: Query<(
+        Entity,
+        &HideOnStateChange<TState>,
+        &TweenBackOnStateChange<TState>,
+    )>,
+    mut state_reader: MessageReader<StateTransitionEvent<TState>>,
 ) {
-    for (e, restore) in restore_q {
+    let msgs: Vec<_> = state_reader.read().collect();
+    if msgs.is_empty() {
+        return;
+    }
+
+    for (e, hide, restore) in restore_q {
+        if !msgs.iter().any(|msg| {
+            if let Some(entered) = msg.entered.as_ref()
+                && hide.restore_on_enter_states.contains(entered)
+            {
+                return true;
+            }
+            if let Some(exited) = msg.exited.as_ref()
+                && hide.restore_on_exit_states.contains(exited)
+            {
+                return true;
+            }
+
+            false
+        }) {
+            return;
+        }
+
         or_continue!(cmd.get_entity(e))
             .try_insert((tween::get_relative_translation_anim(
                 restore.0.truncate(),
